requirements.txt
beautifulsoup4==4.12.3
ibm_cloud_sdk_core==3.20.6
ibm_watson==8.1.0
ibm_watsonx_ai==1.1.8
lxml==5.3.0
openai==1.50.2
Pillow==10.4.0
Requests==2.32.3
streamlit==1.38.0
